getting and setting username works but dispach in dispach
