+++ source/frontend/app/stores/GlobalStore.js
@@ -1,53 +1,20 @@
 import { EventEmitter } from 'events';
 import dispatcher from '../dispatcher';
 
-import idb from 'idb';
-var dbPromise = idb.open('zhawoDB', 1, function(upgradeDB) {
-  if (!upgradeDB.objectStoreNames.contains('users')) {
-    upgradeDB.createObjectStore('users', {
-      keyPath: 'id'
-    });
-  }
-});
-
 class GlobalStore extends EventEmitter {
   constructor() {
     super();
-    this.currentUser = this.getUsernameFromDB() || '';
+    this.currentUser = '';
     this.currentUserType = '';
     this.possibleNames = [];
     this.possibleLoginNames = [];
   }
 
-  getUsernameFromDB() {
-    dbPromise
-      .then(db => {
-        return db
-          .transaction('users')
-          .objectStore('users')
-          .getAll();
-      })
-      .then(allObjs => console.log(allObjs));
-    return '';
-  }
-
-  setUsername(username, type) {
-    this.currentUser = name;
-    this.currentUserType = type;
-    dbPromise.then(db => {
-      const tx = db.transaction('users', 'readwrite');
-      tx.objectStore('users').put({
-        id: username,
-        data: { user: username }
-      });
-      return tx.complete;
-    });
-  }
-
   handleActions(action) {
     switch (action.type) {
       case 'SET_CURRENT_USER':
-        this.setUsername(action.payload.name, action.payload.type);
+        this.currentUser = action.payload.name;
+        this.currentUserType = action.payload.type;
         this.emit('current_user_changed');
         break;
       case 'GET_POSSIBLE_NAMES_OK':
