+++ source/frontend/package.json
@@ -56,6 +56,7 @@
     "file-loader": "^2.0.0",
     "html-webpack-plugin": "^3.2.0",
     "husky": "^1.0.0-rc.14",
+    "idb": "^2.1.3",
     "jest": "^23.6.0",
     "jest-enzyme": "^6.1.2",
     "jest-fetch-mock": "^1.6.6",
